+++ tests/function/complement.js
@@ -5,6 +5,8 @@ test('complement -- Basic functionality', t => {
   t.false(complement(Array.isArray, []))
   t.true(complement(Array.isArray, 10))
   t.same(complement(Number.isInteger, 42), false)
+  t.same(complement(x => x > 10, 11), false)
+  t.same(complement(x => x < 10, 11), true)
   t.end()
 })